fix: set malformed address threshold as environment variable
