+++ functions/raddStoreRegistryLambda/README.md
@@ -13,7 +13,8 @@
 | RADD_STORE_GENERATION_CONFIG_PARAMETER |                         |     no       |
 | RADD_STORE_REGISTRY_API_URL            |                         |     yes      |
 | AWS_LOCATION_REGION                    | eu-central-1            |     yes      |
-| AWS_LOCATION_REQUESTS_PER_SECOND       | 95                      |     no       |
+| AWS_LOCATION_REQUESTS_PER_SECOND       | 95                      |     yes      |
+| MALFORMED_ADDRESS_THRESHOLD            | 0.7                     |     yes      |
 
 ## Istruzioni per la gestione delle configurazioni per il csv dello store locator
